jobfinder/requirements.txt
django>=6.0
djangorestframework>=3.16
django-cors-headers>=4.0
django-mongodb-backend>=6.0
pymongo>=4.7,<5.0
dnspython>=2.0
gunicorn>=21.0
python-decouple>=3.8
Pillow>=10.0
PyJWT>=2.0
